Fix Win32 build errors.

+++ src/Backends/Vulkan/src/tlas.cpp
@@ -113,7 +113,7 @@ void VulkanTopLevelAccelerationStructure::build(const VulkanCommandBuffer& comma
     else if (maxSize < requiredMemory) [[unlikely]]
         throw ArgumentOutOfRangeException("maxSize", 0ull, maxSize, requiredMemory, "The maximum available size is not sufficient to contain the acceleration structure.");
     else if (buffer->size() < offset + requiredMemory) [[unlikely]]
-        throw ArgumentOutOfRangeException("buffer", 0ull, buffer->size(), offset + requiredMemory, "The buffer does not contain enough memory after offset {0} to fully contain the acceleration structure.", offset);
+        throw ArgumentOutOfRangeException("buffer", 0ull, (UInt64)buffer->size(), offset + requiredMemory, "The buffer does not contain enough memory after offset {0} to fully contain the acceleration structure.", offset);
         
     // If the acceleration structure allows for compaction, create a query pool in order to query the compacted size later.
     if (LITEFX_FLAG_IS_SET(m_impl->m_flags, AccelerationStructureFlags::AllowCompaction))
@@ -174,7 +174,7 @@ void VulkanTopLevelAccelerationStructure::update(const VulkanCommandBuffer& comm
     else if (maxSize < requiredMemory) [[unlikely]]
         throw ArgumentOutOfRangeException("maxSize", 0ull, maxSize, requiredMemory, "The maximum available size is not sufficient to contain the acceleration structure.");
     else if (buffer->size() < offset + requiredMemory) [[unlikely]]
-        throw ArgumentOutOfRangeException("buffer", 0ull, buffer->size(), offset + requiredMemory, "The buffer does not contain enough memory after offset {0} to fully contain the acceleration structure.", offset);
+        throw ArgumentOutOfRangeException("buffer", 0ull, (UInt64)buffer->size(), offset + requiredMemory, "The buffer does not contain enough memory after offset {0} to fully contain the acceleration structure.", offset);
 
     // If the acceleration structure allows for compaction, reset the query pool.
     if (LITEFX_FLAG_IS_SET(m_impl->m_flags, AccelerationStructureFlags::AllowCompaction))
@@ -220,7 +220,7 @@ void VulkanTopLevelAccelerationStructure::copy(const VulkanCommandBuffer& comman
     if (buffer == nullptr)
         buffer = destination.m_impl->m_buffer->size() >= requiredMemory[0] ? destination.m_impl->m_buffer : device.factory().createBuffer(BufferType::AccelerationStructure, ResourceHeap::Resource, requiredMemory[0], 1, ResourceUsage::AllowWrite);
     else if (buffer->size() < offset + requiredMemory[0]) [[unlikely]]
-        throw ArgumentOutOfRangeException("buffer", 0ull, buffer->size(), offset + requiredMemory[0], "The buffer does not contain enough memory after offset {0} to fully contain the acceleration structure.", offset);
+        throw ArgumentOutOfRangeException("buffer", 0ull, (UInt64)buffer->size(), offset + requiredMemory[0], "The buffer does not contain enough memory after offset {0} to fully contain the acceleration structure.", offset);
 
     // Create or reset query pool on destination, if required. 
     if (LITEFX_FLAG_IS_SET(destination.flags(), AccelerationStructureFlags::AllowCompaction))
